# MicroServices/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Profile Service
  description: 'Manage profiles'
  version: 0.0.1
  
servers: 
  - url: http://localhost:8080
  
paths:
  /PS/profiles:
    get:
      summary: Returns a list of all users.
      responses: 
        200:
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'
                    
    put: 
      summary: Create a profile
      requestBody:
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        200:
          description: The profile was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        409:
          description: The email adress is already in use
        403: 
          description: The object was not valid
  
  /PS/profiles/{id}:
    parameters:
        - name: id
          in: path
          description: ID of the profile
          required: true
          schema:
            type: integer
    get:
      summary: Return a single profile
      responses:
        200:
          description: The profile
          content:
            application/json:
              schema:
                type: string
                items:
                  $ref: '#/components/schemas/Profile'
    delete:
      summary: Delete a profile
      responses: 
        200:
          description: The profile was deleted

  /PS/profiles/{id}/name:
    parameters:
          - name: id
            in: path
            description: ID of the profile
            required: true
            schema:
              type: integer
    get:
      summary: Return the name of a single profile
      responses:
        200:
          description: The name
          content:
            application/json:
              schema:
                type: string
                items:
                  $ref: '#/components/schemas/Profile'
    put:
      summary: Update the name of the profile
      requestBody:
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        200:
          description: The name was updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  
  /PS/profiles/{id}/email:
    parameters:
          - name: id
            in: path
            description: ID of the profile
            required: true
            schema:
              type: integer
    get:
      summary: Return the email of a single profile
      responses:
        200:
          description: The email
          content:
            application/json:
              schema:
                type: string
                items:
                  $ref: '#/components/schemas/Profile'
    put: 
      summary: Update the email of the profile
      requestBody:
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
      responses:
        200:
          description: The email was updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        409:
          description: The email adress is already in use
        403: 
          description: The email was not valid
          
components: 
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: integer
          example: 4
        name:
          type: string
          example: Theo
        email:
          type: string
          format: email
          example: theo@hotmail.fr
        description: 
          type: string
          example: "A user!"
      required:
          - name
          - email
          